docs: add doc comments where missing in main package
