+++ packages/better-auth/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@kineojs/better-auth",
-  "version": "0.2.3",
+  "version": "0.3.0",
   "description": "Kineo plugin/adapter for Better-Auth.",
   "main": "./dist/index.mjs",
   "types": "./dist/index.d.mts",
