+++ packages/kineo/src/index.ts
@@ -8,7 +8,6 @@ export {
   type Schema,
   type ModelDef,
 } from "./schema";
-export { KineoError, KineoErrorKind } from "./error";
 export { Model, GraphModel } from "./model";
 export type {
   Adapter,
