+++ packages/kineokit/tsconfig.json
@@ -6,5 +6,6 @@
       "@/*": ["src/*"],
       "@tests/*": ["tests/*"]
     }
-  }
+  },
+  "exclude": ["dist"]
 }
